% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_mutate.R
\name{city_mutate}
\alias{city_mutate}
\title{City Mutate}
\usage{
city_mutate(df, pattern = peacePHDs::ukr_cities)
}
\arguments{
\item{df}{A dataframe, containing a column called "translatedText", made by the translaton function}

\item{pattern}{The pattern containing Ukraininan cities}
}
\description{
City Mutate
}
