% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblast_mutate.R
\name{oblast_mutate}
\alias{oblast_mutate}
\title{oblast mutate}
\usage{
oblast_mutate(df, pattern = peacePHDs::ukr_oblast)
}
\arguments{
\item{df}{A dataframe, containing a column called "translatedText", made by the translaton function}

\item{pattern}{The pattern containing Ukraininan oblasts}
}
\description{
oblast mutate
}
