% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_colBind.R
\name{translate_colBind}
\alias{translate_colBind}
\title{translator}
\usage{
translate_colBind(x, text = text)
}
\arguments{
\item{x}{A dataframe with the strings}

\item{text}{the name of the text collumn}
}
\description{
translator
}
