% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukr_minfin_extractor.R
\name{ukr_minfin_extractor}
\alias{ukr_minfin_extractor}
\title{Ukraininan Ministry of Finance Data Extractor}
\usage{
ukr_minfin_extractor(xml_document, css_nr_id = NULL, data_pattern = "'.*'")
}
\arguments{
\item{xml_document}{A graph downloaded from https://index.minfin.com.ua/en/russian-invading/casualties/}

\item{css_nr_id}{marks the date}

\item{data_pattern}{finds the data}
}
\value{
A character string with the nr. of losses/causalties and the date
}
\description{
Ukraininan Ministry of Finance Data Extractor
}
